api/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.17.1
PythonVideoConverter==1.0.3
six==1.16.0
Werkzeug==1.0.1
